shadowcasting implemented

+++ src/constants.rs
@@ -1,7 +1,7 @@
 pub const WINDOW_WIDTH: f32 = 1360.0;
 pub const WINDOW_HEIGHT: f32 = 768.0;
 pub const CHUNK_SIZE: usize = 64;
-pub const LIGHT_RANGE: i32 = 12;
+pub const LIGHT_RANGE: usize = 12;
 pub const PLAYER_SIGHT_RANGE: f32 = 16.0;
 pub const MAX_SOUND_RANGE: f32 = 16.0;
